+++ environments/staging/providers.tf
@@ -6,6 +6,6 @@ provider "aws" {
   region = var.aws_region
 
   default_tags {
-    tags = module.tags.tags
+    tags = local.tags
   }
 }